+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/mapping/ControllerScanner.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet.mvc.mapping;
 
 import static java.util.stream.Collectors.toMap;
